deadlier fireballs, health, and scores.
As time goes on spawning fireballs have a chance of being larger and
faster, this chance of larger and faster fireballs also increases.
Furthermore at the onset there are only 10 fireballs on screen at once,
this number will rise as well.

The player now has health that is displayed as a red bar at the top of
the screen, when hit by fireballs they lose health based on the size and
speed of the fireball.

A few random objects were added to the screen: cobblestones, planters, a
tree.

Whenever a player dies, they start to spin very faster and then explode
leaving only glowing embers behind.

Now as time goes on the player earns points based on how much time has
passed and how much health they still have.